+++ VirtualShelter/src/main/java/hu/bme/aut/vshelter/dal/impl/BreedFacadeJPAImpl.java
@@ -54,7 +54,7 @@ public class BreedFacadeJPAImpl implements BreedFacade {
 	}
 
 	@Override
-	public int getBreedIdfromSpeciesName(String breedName)
+	public long getBreedIdfromSpeciesName(String breedName)
 			throws VirtualShelterException {
 		List<Integer> list = em.createQuery(
 		        "SELECT b.id FROM Breed b WHERE b.breedName = ?1").setParameter(1, breedName).getResultList();
